Refactor dictfetchall, PhonyCursor to helper file
